03-executive-dashboard/requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
